Tidying up files and including uglify-js minification
