+++ _data/scripts/processImages.js
