+++ _includes/foot.html
@@ -15,7 +15,10 @@
 
 {% include custom-foot.html %}
 
-<script src="{{ "/assets/scripts/bundle.js" | absolute_url }}"></script>
+<script type="text/javascript">
+{% include javascript/config.js %}
+</script>
+<script src="{{ "/assets/scripts/bundle.min.js" | absolute_url }}"></script>
 {%- if site.google_maps_api_key -%}
 {%- include google-maps.html -%}
 {%- endif -%}
