+++ _includes/head.html
@@ -6,7 +6,6 @@
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=PT+Sans:wght@400;700&display=swap" rel="stylesheet">
-    <script>var baseurl = '{{ "/" | absolute_url }}';</script>
     <link rel="stylesheet" href="{{ "/assets/css/style.css" | absolute_url }}">
     <meta name="robots" content="noindex">
     {%- if site.google_analytics -%}
