+++ _includes/javascript/map.js
@@ -1,32 +1,54 @@
 /**
- * Google maps API functions for spacefinder
+ * Leafletjs functions for SpaceFinder
  */
+document.addEventListener( 'DOMContentLoaded', () => {
+    initMap();
+});
 
 /**
  * Initialise map and set listeners to set up markers when loaded
- * This function needs to be global as it is used as a callback by
- * the google maps script tag src.
  */
 function initMap() {
-    splog( 'initMap', 'map.js' );
+    splog( 'initMap', 'map-leaflet.js' );
     document.addEventListener( 'sfmaploaded', checkGeo );
     document.addEventListener( 'filtersapplied', filterMarkers );
     document.addEventListener( 'spacesloaded', maybeSetupMap );
+    document.addEventListener( 'filtersloaded', maybeSetupMap );
     document.addEventListener( 'sfmaploaded', maybeSetupMap );
-    spacefinder.map = new google.maps.Map(document.getElementById('map'), {
-        center: spacefinder.currentLoc,
-        zoom: spacefinder.startZoom,
-        mapTypeId: google.maps.MapTypeId.ROADMAP,
+    spacefinder.map = L.map( 'map' ).setView([spacefinder.currentLoc.lat, spacefinder.currentLoc.lng], spacefinder.startZoom );
+    spacefinder.osm = L.tileLayer( 'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
+        maxZoom: 19,
+        attribution: '© <a target="attribution" href="http://www.openstreetmap.org/copyright">OpenStreetMap</a>'
+    }).addTo( spacefinder.map );
+    spacefinder.esri_sat = L.tileLayer( 'https://server.arcgisonline.com/ArcGIS/rest/services/World_Imagery/MapServer/tile/{z}/{y}/{x}', {
+        maxZoom: 19,
+	    attribution: 'Tiles © Esri - Source: Esri, i-cubed, USDA, USGS, AEX, GeoEye, Getmapping, Aerogrid, IGN, IGP, UPR-EGP, and the GIS User Community'
     });
-    google.maps.event.addListenerOnce( spacefinder.map, 'tilesloaded', () => {
-        spacefinder.mapLoaded = true;
-        document.dispatchEvent( new Event( 'sfmaploaded' ) );
-    });
+    spacefinder.map.addControl( new L.Control.Fullscreen( { position: 'topright' } ) );
+    spacefinder.mapLoaded = true;
+    spacefinder.viewdata = {
+        'street': {
+            btnText: 'Street',
+            btnLabel: 'Switch to Street View',
+            btnClass: 'maptype-street',
+            tileLayer: spacefinder.osm
+        },
+        'satellite': {
+            btnText: 'Satellite',
+            btnLabel: 'Switch to Satellite View',
+            btnClass: 'maptype-satellite',
+            tileLayer: spacefinder.esri_sat
+        }
+    };
+
+    document.dispatchEvent( new Event( 'sfmaploaded' ) );
+
     /**
-     * Returns to list view from map "more info" button
+     * Add click listeners
      */
     document.addEventListener( 'click', event => {
-        if ( event.target.classList.contains('show-list') ) {
+        /* Returns to list view from map "more info" button */
+        if ( event.target.classList.contains( 'show-list' ) ) {
             event.preventDefault();
             document.dispatchEvent( new CustomEvent( 'viewchange', {
                 bubbles: true,
@@ -37,6 +59,30 @@ function initMap() {
                 }
             } ) );
         }
+        /* prevents the close button on popups changing the anchor */
+        if ( event.target.classList.contains( 'leaflet-popup-close-button' ) || event.target.parentNode.classList.contains( 'leaflet-popup-close-button' ) ) {
+            event.preventDefault();
+        }
+    });
+
+    /**
+     * make sure the map knows about resizing, both of the main
+     * window, and when a change in view happens (which may reduce
+     * the area taken up by the map component)
+     */
+    window.addEventListener( 'resize', () => {
+        clearTimeout( spacefinder.resizeTimeout );
+        spacefinder.resizeTimeout = setTimeout( () => {
+            splog( 'invalidating map size - resize event', 'map-leaflet.js' );
+            spacefinder.map.invalidateSize( true );
+        }, 200);
+    });
+    document.addEventListener( 'viewchange', () => {
+        splog( 'view changed', 'map-leaflet.js' );
+        window.setTimeout( () => {
+            splog( 'invalidating map size - viewchange event', 'map-leaflet.js' );
+            spacefinder.map.invalidateSize( true );
+        }, 500);
     });
 }
 
@@ -45,43 +91,117 @@ function initMap() {
  * the map is fully loaded and the space data is fully loaded.
  */
 function maybeSetupMap() {
-    splog( 'maybeSetupMap', 'map.js' );
-    if ( spacefinder.mapLoaded && spacefinder.spacesLoaded ) {
-        spacefinder.mapBounds = new google.maps.LatLngBounds();
-        spacefinder.infoWindow = new google.maps.InfoWindow({
-            maxWidth: 350
-        });
-        /* add each spoace to the map using a marker */
+    splog( 'maybeSetupMap', 'map-leaflet.js' );
+    if ( spacefinder.mapLoaded && spacefinder.spacesLoaded && spacefinder.filtersLoaded ) {
+
+        /* collect latLng coordinates here to define map bounds */
+        let pointsArray = [];
+        
+        /**
+         * Initialise marker cluster group
+         * @see https://github.com/Leaflet/Leaflet.markercluster
+         */
+        spacefinder.markergroup = L.markerClusterGroup({
+			disableClusteringAtZoom: 17,
+			zoomToBoundsOnClick: true,
+			spiderfyOnMaxZoom: false,
+			polygonOptions: {
+				color: '#c70000',
+				fillColor: '#c70000'
+			}
+		});
+        
+        /* add each space to the map using a marker */
         for ( let i = 0; i < spacefinder.spaces.length; i++ ) {
             if ( spacefinder.spaces[i].lat && spacefinder.spaces[i].lng ) {
-                var spacePosition = new google.maps.LatLng( spacefinder.spaces[i].lat, spacefinder.spaces[i].lng );
-                spacefinder.spaces[i].marker = new google.maps.Marker({
-                    position: spacePosition,
-                    map: spacefinder.map,
+                var spacePosition = L.latLng( spacefinder.spaces[i].lat, spacefinder.spaces[i].lng );
+                pointsArray.push( [ spacefinder.spaces[i].lat, spacefinder.spaces[i].lng ] );
+                spacefinder.spaces[i].marker = L.marker( spacePosition, {
+                    alt: spacefinder.spaces[i].title,
                     title: spacefinder.spaces[i].title,
-                    optimized: false,
-                    icon: {
-                        path: google.maps.SymbolPath.CIRCLE,//'M0-30.5c-5.7,0-10.2,4.6-10.2,10.2C-10.2-14.6,0,0,0,0s10.2-14.6,10.2-20.2C10.2-25.9,5.7-30.5,0-30.5z M0-17.7c-1.6,0-3-1.3-3-3s1.3-3,3-3s3,1.3,3,3S1.6-17.7,0-17.7z',
-                        fillColor: "#c70000",
-                        strokeColor: "#910000",
-                        fillOpacity: 1,
-                        scale: 8,
-                        strokeWeight: 4
-                    }
+                    icon: getSVGIcon( 'space-marker' )
                 });
-                spacefinder.spaces[i].marker.infoContent = getSpaceInfoWindowContent( spacefinder.spaces[i] );
-                google.maps.event.addListener( spacefinder.spaces[i].marker, 'click', function (e) {
-                    spacefinder.infoWindow.setContent( spacefinder.spaces[i].marker.infoContent );
-                    spacefinder.infoWindow.open( spacefinder.map, spacefinder.spaces[i].marker );
-                    selectSpace(spacefinder.spaces[i].id, 'map');
-                });
-                spacefinder.mapBounds.extend( spacePosition );
+                spacefinder.markergroup.addLayer( spacefinder.spaces[i].marker );
+                /* set the popup for the marker */
+                spacefinder.spaces[i].popup = L.popup().setContent( getSpaceInfoWindowContent( spacefinder.spaces[i] ) );
+                spacefinder.spaces[i].popup.spaceID = spacefinder.spaces[i].id;
+                spacefinder.spaces[i].marker.bindPopup( spacefinder.spaces[i].popup );
             }
         }
-        google.maps.event.addListener(spacefinder.infoWindow,'closeclick',function(){
-            deselectSpaces(false);
+
+        /* add the markers to the map */
+        spacefinder.map.addLayer( spacefinder.markergroup );
+
+        /* use popupopen and popupclose events to select and deselect spaces from map */
+        spacefinder.map.on( 'popupopen', event => {
+            zoomMapToSpace( event.popup.spaceID );
+            document.dispatchEvent( new CustomEvent( 'spaceSelectedOnMap', { bubbles: true, detail: { id: event.popup.spaceID, src: 'map' } } ) );
         });
-        fitAllBounds( spacefinder.mapBounds );
+        spacefinder.map.on( 'popupclose', event => {
+            document.dispatchEvent( new CustomEvent( 'spaceDeselectedFromMap', { bubbles: true, detail: { id: event.popup.spaceID } } ) );
+        });
+
+        /* respond to corresponding events from list */
+        document.addEventListener( 'spaceSelected', event => { zoomMapToSpace( event.detail.id ) } );
+        document.addEventListener( 'spaceDeselected', deselectSpacesFromMap );
+    
+        /* Make sure the map view encompasses all markers */
+        if ( pointsArray.length ) {
+            spacefinder.map.fitBounds( pointsArray );
+        }
+
+        /* save the map bounds and zoom to enable resetting */
+        spacefinder.mapBounds = spacefinder.map.getBounds();
+        spacefinder.mapZoom = parseInt( spacefinder.map.getZoom() );
+        
+        /**
+         * Create a button to switch base layers between streets (OpenStreetMap)
+         * and satellite (ESRI).
+         */
+        L.Control.mapTypeControl = L.Control.extend({
+            onAdd: function(map) {
+                let sd = spacefinder.viewdata.satellite;
+                const mapTypeButton = document.createElement( 'button' );
+                mapTypeButton.innerHTML = sd.btnText;
+                mapTypeButton.classList.add( 'maptype-button' );
+                mapTypeButton.classList.add( sd.btnClass );
+                mapTypeButton.setAttribute( 'aria-label', sd.btnLabel );
+                mapTypeButton.setAttribute( 'title', sd.btnLabel );
+                mapTypeButton.setAttribute( 'data-currentType', 'street' );
+                return mapTypeButton;
+            },
+            onRemove: function(map) {
+                // Nothing to do here
+            }
+        });
+
+        /* constructor */
+        L.control.mapTypeControl = function( opts ) {
+            return new L.Control.mapTypeControl( opts );
+        }
+
+        /* add to map */
+        L.control.mapTypeControl( { position: 'topright' } ).addTo( spacefinder.map );
+
+        /* add listener to map type button to toggle base layer */
+        document.addEventListener('click', event => {
+            if ( event.target.matches( '.maptype-button' ) ) {
+                let currentType = event.target.getAttribute( 'data-currentType' );
+                let newType = currentType == 'street' ? 'satellite': 'street';
+                let mapTypeButton = document.querySelector( 'button.maptype-button' );
+                if ( mapTypeButton ) {
+                    mapTypeButton.classList.replace( spacefinder.viewdata[ newType ].btnClass, spacefinder.viewdata[ currentType ].btnClass );
+                    mapTypeButton.innerHTML = spacefinder.viewdata[ currentType ].btnText;
+                    mapTypeButton.setAttribute( 'aria-label', spacefinder.viewdata[ currentType ].btnLabel );
+                    mapTypeButton.setAttribute( 'title', spacefinder.viewdata[ currentType ].btnLabel );
+                    mapTypeButton.setAttribute( 'data-currentType', newType );
+                    spacefinder.viewdata[currentType].tileLayer.removeFrom( spacefinder.map );
+                    spacefinder.viewdata[newType].tileLayer.addTo( spacefinder.map );
+                }
+            }
+        });
+        
+        /* let eveyone know we are ready */
         document.dispatchEvent( new Event( 'sfmapready' ) );
     }
 }
@@ -92,71 +212,81 @@ function maybeSetupMap() {
  * @returns {String} HTML content for space infoWindow
  */
 function getSpaceInfoWindowContent( space ) {
-    splog( 'getSpaceInfoWindowContent', 'map.js' );
-    return '<div class="spaceInfoWindow"><h3>'+space.title+'</h3><p>'+space.description+'</p><button class="show-list">More info&hellip;</button></div>';
+	let info = [];
+	info.push( space.space_type );
+	if ( space.floor !== '' ) {
+		info.push( space.floor );
+	}
+	if ( space.building !== '' ) {
+		info.push( space.building );
+	}
+	let content = '<div class="spaceInfoWindow"><h3>'+space.title+'</h3>';
+	content += '<p class="info">' + info.join(', ') + '</p>';
+	content += '<p class="description">' + space.description + '</p>';
+	content += '<button class="show-list">More info&hellip;</button></div>';
+	return content;
 }
-
 
 /**
- * Takes the current bounds of the map and ensures all
- * markers fit within the visible area while keeping the
- * map centred on a given location (by setting the zoom).
- * @param {google.maps.LatLngBounds} b bounds object
+ * Returns an object to be used in the map to make a leaflet icon
+ * @param {String} className CSS class to be used on the icon
+ * @return {Object}
  */
-function fitAllBounds(b) {
-    splog( 'fitAllBounds', 'map.js' );
-    // Get north east and south west markers bounds coordinates
-    var ne = b.getNorthEast();
-    var sw = b.getSouthWest();
-    // Get the current map bounds
-    var mapBounds = spacefinder.map.getBounds();
-    // Check if map bounds contains both north east and south west points
-    if (mapBounds.contains(ne) && mapBounds.contains(sw)) {
-        // Everything fits
-        return;
-    } else {
-        var mapZoom = spacefinder.map.getZoom();
-        if (mapZoom > 0) {
-            // Zoom out
-            spacefinder.map.setZoom(mapZoom - 1);
-            // Try again
-            fitAllBounds(b);
-        }
-    }
+function getSVGIcon( c ) {
+	return L.divIcon({
+  		html: `<svg width="32" height="32" viewBox="0 0 32 32" version="1.1" xmlns="http://www.w3.org/2000/svg"><circle cx="16" cy="16" r="10" stroke-width="6"></circle></svg>`,
+		className: c,
+  		iconSize: [32, 32],
+  		iconAnchor: [16, 16]
+	});
 }
 
+/**
+ * Re-centres map
+ * Not currently used (written to enable recentring like in a google maps <recentre> button)
+ */
 function recentreMap() {
-    splog( 'recentreMap', 'map.js' );
+    splog( 'recentreMap', 'map-leaflet.js' );
     let newCenter = geolocationActive() ? spacefinder.personLoc: spacefinder.currentLoc;
-    spacefinder.map.setCenter( newCenter );
-    fitAllBounds( spacefinder.mapBounds );
+    spacefinder.map.panTo( newCenter );
+    spacefinder.map.fitBounds( spacefinder.mapBounds );
 }
+
 /**
  * Zooms the map to show a particular space
  * @param {Object} space
  */
-function zoomMapToSpace( space ) {
-    splog( 'zoomMapToSpace', 'map.js' );
-    let newCenter = new google.maps.LatLng( space.lat, space.lng );
-    spacefinder.map.panTo( newCenter );
-    spacefinder.map.setZoom(18);
-    spacefinder.infoWindow.setContent( getSpaceInfoWindowContent( space ) );
-    spacefinder.infoWindow.open( spacefinder.map, space.marker );    
+function zoomMapToSpace( spaceid ) {
+    splog( 'zoomMapToSpace', 'map-leaflet.js' );
+    let space = getSpaceById( spaceid );
+    let newCenter = L.latLng( space.lat, space.lng );
+    space.popup.setLatLng( newCenter ).openOn( spacefinder.map );
+    spacefinder.map.setView( newCenter, 18 );
+}
+
+
+/**
+ * Resets the map after a space has been selected
+ */
+ function deselectSpacesFromMap() {
+    splog( 'deselectSpacesFromMap', 'map-leaflet.js' );
+    spacefinder.map.closePopup();
 }
 
 /**
  * Filters the markers on the map
  */
 function filterMarkers() {
-    splog( 'filterMarkers', 'map.js' );
-    document.querySelectorAll('.list-space').forEach( el => {
-        let space = getSpaceById( el.getAttribute('data-id') );
-        if ( el.classList.contains('hidden') ) {
-            space.marker.setMap( null );
-        } else {
-            space.marker.setMap( spacefinder.map );
+    splog( 'filterMarkers', 'map-leaflet.js' );
+    let markersToAdd = [];
+    document.querySelectorAll( '.list-space' ).forEach( element => {
+        let space = getSpaceById( element.getAttribute( 'data-id' ) );
+        if ( ! element.classList.contains( 'hidden' ) ) {
+            markersToAdd.push( space.marker );
         }
     });
+    spacefinder.markergroup.clearLayers();
+    spacefinder.markergroup.addLayers( markersToAdd );
 }
 
 /*******************************************************************
@@ -168,11 +298,11 @@ function filterMarkers() {
  * @param {boolean} enable which way to toggle
  */
 function toggleGeolocation( enable ) {
-    splog( 'toggleGeolocation', 'map.js' );
+    splog( 'toggleGeolocation', 'map-leaflet.js' );
     if ( enable ) {
-        document.querySelectorAll( '.geo-button' ).forEach( e => e.disabled = false );
+        document.querySelectorAll( '.geo-button' ).forEach( element => element.disabled = false );
     } else {
-        document.querySelectorAll( '.geo-button' ).forEach( e => e.disabled = true );
+        document.querySelectorAll( '.geo-button' ).forEach( element => element.disabled = true );
     }
 }
 
@@ -183,29 +313,29 @@ function toggleGeolocation( enable ) {
  * @param {boolean} activate which way to toggle
  */
 function activateGeolocation( activate ) {
-    splog( 'activateGeolocation', 'map.js' );
+    splog( 'activateGeolocation', 'map-leaflet.js' );
     if ( activate ) {
-        document.querySelectorAll( '.geo-button' ).forEach( e => {
-            e.classList.add('active');
-            e.setAttribute('aria-label','Stop using my location')
-            e.setAttribute('title','Stop using my location')
+        document.querySelectorAll( '.geo-button' ).forEach( element => {
+            element.classList.add( 'active' );
+            element.setAttribute( 'aria-label', 'Stop using my location' );
+            element.setAttribute( 'title', 'Stop using my location' );
         });
         document.addEventListener( 'userlocationchanged', movePersonMarker );
-        document.dispatchEvent(new CustomEvent('sfanalytics', {
+        document.dispatchEvent(new CustomEvent( 'sfanalytics', {
             detail: {
                 type: 'geostart'
             }
         }));
     } else {
-        document.querySelectorAll( '.geo-button' ).forEach( e => {
-            e.classList.remove('active');
-            e.setAttribute('aria-label','Use my location')
-            e.setAttribute('title','Use my location')
+        document.querySelectorAll( '.geo-button' ).forEach( element => {
+            element.classList.remove( 'active' );
+            element.setAttribute( 'aria-label', 'Use my location' );
+            element.setAttribute( 'title', 'Use my location' );
         });
         document.removeEventListener( 'userlocationchanged', movePersonMarker );
         /* remove sorting indicator from all buttons */
-        document.getElementById( 'sortdistance' ).setAttribute('data-sortdir', '');
-        document.dispatchEvent(new CustomEvent('sfanalytics', {
+        document.getElementById( 'sortdistance' ).setAttribute( 'data-sortdir', '' );
+        document.dispatchEvent(new CustomEvent( 'sfanalytics', {
             detail: {
                 type: 'geoend'
             }
@@ -225,7 +355,7 @@ function activateGeolocation( activate ) {
  * @see getUserPosition()
  */
 function movePersonMarker() {
-    splog( 'movePersonMarker', 'map.js' );
+    splog( 'movePersonMarker', 'map-leaflet.js' );
     /* move person marker */
     if ( spacefinder.personMarker ) {
         spacefinder.personMarker.setPosition( spacefinder.personLoc );
@@ -233,16 +363,16 @@ function movePersonMarker() {
     /* update distances to each space */
     updateDistances();
     /* see if the spaces are sorted by distance */
-    let btn = document.querySelector('#sortdistance[data-sortdir$="sc"');
+    let btn = document.querySelector( '#sortdistance[data-sortdir$="sc"' );
     if ( btn !== null ) {
         /* determine direction from current attribute value */
-        let sortdir = document.getElementById('sortdistance').getAttribute('data-sortdir');
+        let sortdir = document.getElementById( 'sortdistance' ).getAttribute( 'data-sortdir' );
         let dir = ( sortdir == 'desc' ) ? false: true;
         /* re-sort spaces */
         sortSpaces( 'sortdistance', dir );
     }
     /* centre the map on the person */
-    spacefinder.map.setCenter( spacefinder.personLoc );
+    spacefinder.map.panTo( spacefinder.personLoc );
 }
 
 /**
@@ -250,7 +380,7 @@ function movePersonMarker() {
  * @returns {boolean}
  */
 function geolocationEnabled() {
-    splog( 'geolocationEnabled', 'map.js' );
+    splog( 'geolocationEnabled', 'map-leaflet.js' );
     const btn = document.querySelector( '.geo-button' );
     if ( btn !== null ) {
         return btn.disabled == false;
@@ -263,7 +393,7 @@ function geolocationEnabled() {
  * @returns {boolean}
  */
 function geolocationActive() {
-    splog( 'geolocationActive', 'map.js' );
+    splog( 'geolocationActive', 'map-leaflet.js' );
     return ( document.querySelector( '.geo-button.active' ) !== null ? true: false );
 }
 
@@ -271,7 +401,7 @@ function geolocationActive() {
  * Performs checks for geolocation permissions and services when the map has loaded
  */
 function checkGeo() {
-    splog( 'checkGeo', 'map.js' );
+    splog( 'checkGeo', 'map-leaflet.js' );
     /* first see if geolocation is available on the device */
     checkGeoAvailable();
     /* check to see if it is enabled to determine initial button states */
@@ -284,7 +414,7 @@ function checkGeo() {
  * watches for updates to permissions.
  */
 function checkGeoPermissions() {
-    splog( 'checkGeoPermissions', 'map.js' );
+    splog( 'checkGeoPermissions', 'map-leaflet.js' );
     /* check for permissions query */
     if ( 'permissions' in navigator && navigator.permissions.query ) {
         /* query geolocation permissions */
@@ -317,20 +447,29 @@ function checkGeoPermissions() {
  * adds buttons to activate it and adds listeners to buttons.
  */
 function checkGeoAvailable() {
-    splog( 'checkGeoAvailable', 'map.js' );
+    splog( 'checkGeoAvailable', 'map-leaflet.js' );
     if ( 'geolocation' in navigator ) {
         /* make button for map to let user activate geolocation */
-        const locationButton = document.createElement( 'button' );
-        locationButton.innerHTML = '';
-        locationButton.classList.add('geo-button');
-        locationButton.classList.add('icon-my-location');
-        locationButton.setAttribute('aria-label', 'Use my location');
-        locationButton.setAttribute('title', 'Use my location');
-        spacefinder.map.controls[google.maps.ControlPosition.RIGHT_TOP].push(locationButton);
+        L.Control.geoControl = L.Control.extend({
+            onAdd: function(map) {
+                const locationButton = document.createElement( 'button' );
+                locationButton.innerHTML = '';
+                locationButton.classList.add( 'geo-button' );
+                locationButton.classList.add( 'icon-my-location' );
+                locationButton.setAttribute( 'aria-label', 'Use my location' );
+                locationButton.setAttribute( 'title', 'Use my location' );
+                return locationButton;
+            },
+            onRemove: function(map) {}
+        });
+        L.control.geoControl = function(opts) {
+            return new L.Control.geoControl(opts);
+        }
+        L.control.geoControl( { position: 'topright' } ).addTo( spacefinder.map );
 
         /* add listener to buttons to toggle geolocation */
-        document.addEventListener('click', e => {
-            if ( e.target.matches('.geo-button') ) {
+        document.addEventListener( 'click', event => {
+            if ( event.target.matches( '.geo-button' ) ) {
                 if ( ! geolocationEnabled() ) {
                     return;
                 }
@@ -355,15 +494,15 @@ function checkGeoAvailable() {
  * and deactivates geolocation controls.
  */
 function forgetUserPosition() {
-    splog( 'forgetUserPosition', 'map.js' );
+    splog( 'forgetUserPosition', 'map-leaflet.js' );
     /* stop watching user position */
     navigator.geolocation.clearWatch( spacefinder.watchID );
     /* remove person marker from map */
-    spacefinder.personMarker.setMap(null);
+    spacefinder.personMarker.remove();
     /* make location buttons inactive */
     activateGeolocation( false );
     /* re-centre map */
-    spacefinder.map.setCenter( spacefinder.currentLoc );
+    spacefinder.map.panTo( spacefinder.currentLoc );
 }
 /**
  * Gets the current position of the user device, centres the
@@ -375,32 +514,20 @@ function forgetUserPosition() {
  * button to recentre? (but not moving the marker)
  */
 function getUserPosition() {
-    splog( 'getUserPosition', 'map.js' );
+    splog( 'getUserPosition', 'map-leaflet.js' );
 	navigator.geolocation.getCurrentPosition( position => {
         /* centre the map on the user coordinates */
 		spacefinder.personLoc.lat = position.coords.latitude;
 		spacefinder.personLoc.lng = position.coords.longitude;
         if ( ! spacefinder.mapBounds.contains( spacefinder.personLoc ) ) {
             toggleGeolocation( false );
-            openAlertDialog('Sorry...', 'You need to be a bit nearer to use this feature.');
+            openAlertDialog( 'Sorry...', 'You need to be a bit nearer to use this feature.' );
             return;
         }
         /* centre the map on the user position */
-		spacefinder.map.setCenter( spacefinder.personLoc );
+		spacefinder.map.panTo( spacefinder.personLoc );
         /* add a person marker */
-		spacefinder.personMarker = new google.maps.Marker({
-			position: spacefinder.personLoc,
-			map: spacefinder.map,
-			title: 'Your location',
-            icon: {
-                path: google.maps.SymbolPath.CIRCLE,
-                scale: 10,
-                fillOpacity: 1,
-                strokeWeight: 2,
-                fillColor: '#003a52',
-                strokeColor: '#ffffff',
-            },
-		});
+		spacefinder.personMarker = L.marker( spacefinder.personLoc, { alt: 'Your location' } ).addTo( spacefinder.map );
         activateGeolocation( true );
         /* watch for changes in the user position and update the map by firing an event */
 		spacefinder.watchID = navigator.geolocation.watchPosition( position => {
@@ -434,16 +561,16 @@ function getUserPosition() {
  * to the user position.
  */
 function updateDistances() {
-    splog( 'updateDistances', 'map.js' );
+    splog( 'updateDistances', 'map-leaflet.js' );
     if ( geolocationActive() ) {
         spacefinder.spaces.forEach( (space, index) => {
             spacefinder.spaces[index].distancefromcentre = haversine_distance( spacefinder.personLoc, { lat: space.lat, lng: space.lng } );
-            document.querySelector('[data-id="' + space.id + '"]').setAttribute('data-sortdistance', spacefinder.spaces[index].distancefromcentre );
+            document.querySelector( '[data-id="' + space.id + '"]').setAttribute( 'data-sortdistance', spacefinder.spaces[index].distancefromcentre );
         });
     } else {
-        let spacenodes = document.querySelectorAll('.list-space');
+        let spacenodes = document.querySelectorAll( '.list-space' );
         if ( spacenodes !== null ) {
-            spacenodes.forEach( el => el.setAttribute('data-sortdistance', '') );
+            spacenodes.forEach( element => element.setAttribute( 'data-sortdistance', '' ) );
         }
     }
 }
