+++ package.json
@@ -5,16 +5,10 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "clean": "node _scripts/cleanJSON.js",
-    "checkFilters": "node _scripts/checkFilters.js",
     "concat": "node _scripts/concat.js",
-    "cmap": "node _scripts/campusmap.js",
-    "get_pids": "node _scripts/getPlaceIDs.js",
-    "clean_pids": "node _scripts/cleanPlaceIDs.js",
-    "process": "node _scripts/processJSON.js",
-    "resize": "node _scripts/processImages.js",
     "updateOpening": "node _scripts/updateLibraryOpening.js",
-    "crontab": "node _scripts/cron.js"
+    "crontab": "node _scripts/cron.js",
+    "minify": "node _scripts/minify.js"
   },
   "repository": {
     "type": "git",
@@ -24,17 +18,6 @@
   "author": "",
   "license": "Apache-2.0",
   "devDependencies": {
-    "@babel/core": "^7.17.8",
-    "@babel/preset-env": "^7.16.11",
-    "@what3words/api": "^4.0.6",
-    "axios": "^0.26.1",
-    "babel-loader": "^8.2.3",
-    "cross-fetch": "^3.1.5",
-    "http": "^0.0.1-security",
-    "isomorphic-unfetch": "^3.1.0",
-    "sharp": "^0.27.2",
-    "webpack": "^5.70.0",
-    "webpack-cli": "^4.9.2",
-    "yaml": "^2.1.1"
+    "uglify-js": "^3.17.0"
   }
 }
